requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-crispy-forms==1.7.2
django-datatables-view==1.16.0
django-pyodbc==1.1.3
django-pyodbc-azure==2.0.6.1
idna==2.7
oauthlib==2.1.0
pyodbc==4.0.23
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
